+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 
+##[0.6.0] - 2017-04-15
+### Added
+- Camera rotates around watchee on mouse button 0 down
+- Camera pans up/down around on axis perpendicular to sight line on mouse button 0 down.
+
 ## [0.5.0] - 2017-04-14
 ### Added
 - Camera follows flock in web build mode
@@ -18,6 +23,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 - Flier changed to bouncing red pill.
 
-## [0.1.0] 2016-12-**
+## [0.1.0] - 2016-12-**
 ### Added
 - Keyboard controlled flier for flock to follow.